==> scripts/01_preprocessData.Rmd <==
---
title: "Preprocessing"
output:
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    code_folding: 'hide'
    self_contained: true
---


```{r include=FALSE}

library(tidyverse)
library(labelled)
library(knitr)
library(visdat)
library(scales)
library(kableExtra)
theme_set(theme_bw())

```

```{r include=FALSE}
d <- read_rds("../../Data/LQPdata_checkedAndCleaned_v06.rds")
glimpse(d)
```



## Residence security
Person_Rechte
Ranking: CH <- C/Ci <- BEU <- BnonEU <- LEU <- LnonEU <- F <- N 
NAs: S G

https://ma.zh.ch/internet/sicherheitsdirektion/migrationsamt/de/einreise_aufenthalt/ausweise_bewilligungsarten/_jcr_content/contentPar/downloadlist/downloaditems/_bersicht_aufenthalt.spooler.download.1513866469408.pdf/Aufenthaltskategorien+in+der+Schweiz_IW.pdf


```{r echo=TRUE}
d <- d %>% 
  mutate(Person_Aufenhaltssicherheit = case_when(
    Person_Aufenthaltsstat %in% 
      c("Ausweis S (für Schutzbedürftige)", 
        "Ausweis G (Grenzgängerbewilligung)") ~ 
      NA_character_,
    Person_Aufenthaltsstat == "Ausweis F (Vorläufig aufgenommen)" ~ "Ausweis F",
    Person_Aufenthaltsstat == "Ausweis N (für Asylsuchende)" ~ "Ausweis N",
    Person_Aufenthaltsstat == 
      "Ausweis L (Kurzaufenthaltsbewilligung)" & Person_EUBuerger == "Nein" ~ 
      "Ausweis L",
    Person_Aufenthaltsstat == 
      "Ausweis L (Kurzaufenthaltsbewilligung)" & Person_EUBuerger == "Ja" ~ 
      "Ausweis B EU",
    Person_Aufenthaltsstat == 
      "Ausweis B (Aufenthaltsbewilligung)" & Person_EUBuerger == "Nein" ~ 
      "Ausweis B",
    Person_Aufenthaltsstat == 
      "Ausweis B (Aufenthaltsbewilligung)" & Person_EUBuerger == "Ja" ~ 
      "Ausweis B EU",
    Person_Aufenthaltsstat == "Ausweis C (Niederlassungsbewilligung)" ~ 
      "Ausweis C",
    Person_CHpass == "Ja" ~ "CH")) %>% 
  mutate(Person_Aufenhaltssicherheit = factor(Person_Aufenhaltssicherheit, levels = c("CH", "Ausweis C", "Ausweis B EU", "Ausweis B", "Ausweis L", "Ausweis F"), ordered = T))
var_label(d$Person_Aufenhaltssicherheit) <- "Aufenthaltsstatus der Person"

```

```{r echo=TRUE}
d %>% select(Person_CHpass, Person_EUBuerger, Person_Aufenthaltsstat, 
             Person_Aufenhaltssicherheit)
```



## Age

```{r echo=TRUE}
d <- d %>% 
  mutate(Person_Alter = 2019 - Person_Jahrgang)
var_label(d$Person_Alter) <- "Alter der Person (in Jahren)"

```
```{r echo=TRUE}
d %>% select(Person_Jahrgang, Person_Alter)
```



## Single houshold
```{r echo=TRUE}
d <- 
  d %>% mutate(
    Person_Einzelhaushalt = if_else(condition = Person_Haushaltsgr == 1, 
                                    true = "Ja", 
                                    false = "Nein"), 
    Person_Einzelhaushalt = factor(Person_Einzelhaushalt, 
                                   levels = c("Ja", "Nein")),
  )
var_label(d$Person_Einzelhaushalt) <- "Lebt die Person in einem Einzelhaushalt?"
```
```{r}
d$Person_Einzelhaushalt %>% levels()
```


## Level of education
```{r echo=TRUE}
d <- 
  d %>% mutate(
    Person_BildAbschluss = factor(x = Person_BildAbschluss,
                                  levels = c("Obligatorische Schulzeit",
                                             "Berufliche Grundbildung",
                                             "Allgemeinbildende Schule",
                                             "Höhere Berufsbildung",
                                             "Bachelor an Hochschule",
                                             "Master an Hochschule",
                                             "Doktorat"),
                                  ordered = T))
var_label(d$Person_BildAbschluss) <- "Höchster Bildungsabschluss der Person"
```
```{r echo=TRUE}
d %>% select(Person_BildAbschluss)
```


## Houshold size
Delete entries where the number of children is equalt or above the total number
of houshold members

```{r}
d <- d %>% 
  mutate(Person_Haushaltsgr = 
           ifelse(test = Person_Haushaltsgr > Person_KinderU18, 
                   yes = Person_Haushaltsgr, 
                   no = NA)) %>% 
    mutate(Person_Haushaltsgr = 
           ifelse(test = Person_Haushaltsgr != 0, 
                   yes = Person_Haushaltsgr, 
                   no = NA))  
```






## Income per capita

https://www.pewsocialtrends.org/2011/10/03/appendix-b-adjusting-household-income-for-household-size/
http://www.oecd.org/els/soc/OECD-Note-EquivalenceScales.pdf

-> Ask Christop B 

```{r echo=TRUE}
d <- d %>% mutate(
  Person_HaushaltsgrErw = Person_Haushaltsgr - Person_KinderU18,
  Person_HaushaEinkommen =  factor(Person_HaushaEinkommen, 
                                   c("Bis 2'000 Franken", 
                                     "2'001 bis 4'000 Franken", 
                                     "4'001 bis 6'000 Franken", 
                                     "6'001 bis 8'000 Franken", 
                                     "8'001 bis 10'000 Franken" ,
                                     "10'001 bis 12'000 Franken", 
                                     "12'001 bis 14'000 Franken", 
                                     "Mehr als 14'000 Franken")), 
  Person_HaushaEinkommen = recode(Person_HaushaEinkommen,
                                  "Bis 2'000 Franken" = "1000", 
                                  "2'001 bis 4'000 Franken" = "3000", 
                                  "4'001 bis 6'000 Franken" = "5000", 
                                  "6'001 bis 8'000 Franken" = "7000", 
                                  "8'001 bis 10'000 Franken" = "9000",
                                  "10'001 bis 12'000 Franken" = "11000", 
                                  "12'001 bis 14'000 Franken" = "13000", 
                                  "Mehr als 14'000 Franken" = "15000"), 
  Person_HaushaEinkommen = as.integer(as.character(Person_HaushaEinkommen)),
  Person_persEinkommen = round(Person_HaushaEinkommen / Person_HaushaltsgrErw, digits = 0)
)
var_label(d$Person_persEinkommen) <- "Höhe der Einkommens pro Monat in CHF"

```
```{r echo=TRUE}
d %>% select(Person_Haushaltsgr, Person_KinderU18, Person_HaushaEinkommen, 
             Person_persEinkommen)
```


## Workload
```{r}
d %>% 
  select(Person_StundenEngag, Person_StundenPflege, Person_StundenHaush, 
         Person_StundenErwerb) %>% 
  summary()
```


## Data reduction

```{r echo=TRUE}
d2 <- d %>% 
  select(
    Zufriedenheit_allgLeben,
    Meta_Studienregion, Person_Alter, Person_Gender, 
    Person_Aufenhaltssicherheit, Person_Kinder, Person_Einzelhaushalt, 
    Person_BildAbschluss, Person_persEinkommen, Person_Taetigkeitsbereich)
d2 %>% glimpse()
```
```{r}
write_rds(d2, "../../Data/01_preprocessed.rds")
```
